Make mypy stricter and type annotate the rest of the library
There were a few small dormant bugs found via this process (yay!) but it
should mostly not have any runtime effect.

+++ baseplate/lib/file_watcher.py
@@ -96,7 +96,7 @@ class FileWatcher(Generic[T]):
         self,
         path: str,
         parser: Callable[[IO], T],
-        timeout: Optional[int] = None,
+        timeout: Optional[float] = None,
         binary: bool = False,
         encoding: Optional[str] = None,
         newline: Optional[str] = None,
